[ADD] Pricer class from what will inherit concrete pricers
